New deploy settings
This broke SO hard! I upgraded to VS2022 for Mac and it doesn't seem to use Web Deploy anymore, just zip deploy, but it wouldn't do it. Just gave an InternalServerError from the Azure side. I tried 1000 things and nothing worked. I finally downgraded back to VS2019 and it still didn't work. I finally added an SCM setting in Azure (googling ftw!) and it worked. I'll have to look at this again later.